requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.4.1
Markdown==3.4.3
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
zipp==3.15.0
